New Hazard, Changed Background, Comments, & A Jump
Added a new hazard (stabbing hazard) that gives a warning to the player before it quickly enters the screen and retreats. Changed the screen display to 4:3 and made it so white now has access to the entire screen. Added some comments to both player scripts for easier understanding. Applied some physics and gave player1 a simple jump.